Conflicts:
	tita-business/src/main/java/at/ac/tuwien/ifs/tita/business/service/time/IEffortService.java
	tita-dao/src/main/java/at/ac/tuwien/ifs/tita/dao/interfaces/IEffortDao.java
	tita-dao/src/main/java/at/ac/tuwien/ifs/tita/dao/time/EffortDao.java
	tita-dao/src/sql/cre_tables.sql
	tita-dao/src/sql/ins_base_data.sql
	tita-entity/src/main/java/at/ac/tuwien/ifs/tita/entity/Effort.java

+++ tita-business/src/main/java/at/ac/tuwien/ifs/tita/business/service/time/EffortService.java
@@ -78,7 +78,7 @@ public class EffortService implements IEffortService {
     /**
      * {@inheritDoc}
      */
-    public List<Double> getEffortsYears() {
+    public List<Integer> getEffortsYears() {
         return this.timeEffortDao.getTimeEffortsYears();
     }
 
@@ -94,18 +94,16 @@ public class EffortService implements IEffortService {
      * {@inheritDoc}
      */
     @Override
-    public List<UserProjectEffort> getEffortsSummaryForProjectAndUserNames(
-            List<String> projects, List<String> usernames, String grouping) {
-        return timeEffortDao.findEffortsForTiTAProjectAndTimeConsumerId(projects, 
-                                                                        usernames, grouping);
+    public List<UserProjectEffort> getEffortsSummaryForProjectAndUserNames(List<String> projects,
+            List<String> usernames, String grouping) {
+        return timeEffortDao.findEffortsForTiTAProjectAndTimeConsumerId(projects, usernames, grouping);
     }
 
     /**
      * {@inheritDoc}
      */
     @Override
-    public List<UserProjectEffort> getEffortsSummaryForProjectNames(
-            List<String> projects, String grouping) {
+    public List<UserProjectEffort> getEffortsSummaryForProjectNames(List<String> projects, String grouping) {
         return timeEffortDao.findEffortsForTiTAProjectId(projects, grouping);
     }
